LayoutTests/http/tests/xmlhttprequest/xhr-large-document-expected.txt

Tests that XMLHttpRequest.document is valid for large HTML documents

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS xhr.response is not null
PASS successfullyParsed is true

TEST COMPLETE
